backend/requirements.txt
fastapi
uvicorn
torch
torchvision
albumentations
python-multipart
Pillow
